fix(package): refacktor rename some package and add technical skill endpoint
