--- src/main/java/ogya/lokakarya/be/dto/technicalSkill/CreateTechnicalSkill.java
@@ -1,21 +1,0 @@
-package ogya.lokakarya.be.dto.technicalSkill;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.*;
-
-import java.util.Date;
-import java.util.UUID;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Data
-@ToString
-public class CreateTechnicalSkill {
-
-//    @JsonProperty("technical_skill")
-//    private String technicalSKill;
-//
-//    @JsonProperty("enabled")
-//    private boolean enabled=true;
-}
