--- src/main/java/ogya/lokakarya/be/dto/technicalSkill/TechnicalSkillDto.java
@@ -1,31 +1,0 @@
-package ogya.lokakarya.be.dto.technicalSkill;
-
-import lombok.*;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Data
-@ToString
-public class TechnicalSkillDto {
-//    @JsonProperty("id")
-//    private UUID id;
-//
-//    @JsonProperty("technical_skill")
-//    private String technicalSKill;
-//
-//    @JsonProperty("enabled")
-//    private boolean enabled=true;
-//
-//    @JsonProperty("created_at")
-//    private Date createdAt = new Date();
-//
-//    @JsonProperty("created_by")
-//    private UUID createdBy;
-//
-//    @JsonProperty("updated_at")
-//    private Date updatedAt;
-//
-//    @JsonProperty("updated_by")
-//    private UUID updatedBy;
-}
