+++ src/main/java/ogya/lokakarya/be/repository/devplan/DevPlanRepository.java
@@ -1,0 +1,8 @@
+package ogya.lokakarya.be.repository.devplan;
+
+import ogya.lokakarya.be.entity.DevPlan;
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.UUID;
+
+public interface DevPlanRepository extends JpaRepository<DevPlan, UUID> {
+}
