--- src/main/java/ogya/lokakarya/be/repository/groupAttitudeSkill/GroupAttitudeSkillRepository.java
@@ -1,9 +1,0 @@
-package ogya.lokakarya.be.repository.groupAttitudeSkill;
-
-import ogya.lokakarya.be.entity.GroupAttitudeSkill;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface GroupAttitudeSkillRepository extends JpaRepository<GroupAttitudeSkill, UUID> {
-}
