+++ src/main/java/ogya/lokakarya/be/repository/groupattitudeskill/GroupAttitudeSkillRepository.java
@@ -1,0 +1,9 @@
+package ogya.lokakarya.be.repository.groupattitudeskill;
+
+import ogya.lokakarya.be.entity.GroupAttitudeSkill;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface GroupAttitudeSkillRepository extends JpaRepository<GroupAttitudeSkill, UUID> {
+}
