+++ src/main/java/ogya/lokakarya/be/service/AttitudeSkillService.java
@@ -1,11 +1,8 @@
 package ogya.lokakarya.be.service;
 
-import ogya.lokakarya.be.dto.attitudeSkill.AttitudeSkillDto;
-import ogya.lokakarya.be.dto.attitudeSkill.CreateAttitudeSkill;
-import ogya.lokakarya.be.dto.division.CreateDivision;
-import ogya.lokakarya.be.dto.division.DivisionDto;
+import ogya.lokakarya.be.dto.attitudeskill.AttitudeSkillDto;
+import ogya.lokakarya.be.dto.attitudeskill.CreateAttitudeSkill;
 import ogya.lokakarya.be.entity.AttitudeSkill;
-import ogya.lokakarya.be.entity.Division;
 
 import java.util.List;
 import java.util.UUID;
