+++ src/main/java/ogya/lokakarya/be/service/DevPlanService.java
@@ -1,11 +1,9 @@
 package ogya.lokakarya.be.service;
 
-import ogya.lokakarya.be.dto.devPlan.CreateDevPlan;
-import ogya.lokakarya.be.dto.devPlan.DevPlanDto;
-import ogya.lokakarya.be.dto.division.CreateDivision;
-import ogya.lokakarya.be.dto.division.DivisionDto;
+import ogya.lokakarya.be.dto.devplan.CreateDevPlan;
+import ogya.lokakarya.be.dto.devplan.DevPlanDto;
 import ogya.lokakarya.be.entity.DevPlan;
-import ogya.lokakarya.be.entity.Division;
+
 import java.util.List;
 import java.util.UUID;
